Refactor code for the requested changes v1
