+++ assets/css/styles.css
@@ -30,14 +30,15 @@ main {
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
-  align-content: space-evenly;
+  -ms-flex-line-pack: space-evenly;
+      align-content: space-evenly;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   -webkit-box-orient: vertical;
   -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
+      -ms-flex-direction: column;
+          flex-direction: column;
   overflow: hidden;
 }
 
@@ -60,19 +61,18 @@ main {
 
 .messages-container .information {
   -webkit-animation: border-animation 5s;
-  animation: border-animation 5s;
+          animation: border-animation 5s;
 }
 
 .messages-container .winner {
   -webkit-animation: border-winning 10s;
-  animation: border-winning 10s;
+          animation: border-winning 10s;
 }
 
 @-webkit-keyframes border-animation {
   0% {
     border: solid red 1px;
   }
-
   100% {
     border: solid white 1px;
   }
@@ -82,7 +82,6 @@ main {
   0% {
     border: solid red 1px;
   }
-
   100% {
     border: solid white 1px;
   }
@@ -92,7 +91,6 @@ main {
   0% {
     border: solid green 3px;
   }
-
   100% {
     border: solid white 1px;
   }
@@ -102,7 +100,6 @@ main {
   0% {
     border: solid green 3px;
   }
-
   100% {
     border: solid white 1px;
   }
@@ -118,18 +115,18 @@ main {
   display: flex;
   -webkit-box-orient: horizontal;
   -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
+      -ms-flex-direction: row;
+          flex-direction: row;
   -ms-flex-wrap: nowrap;
-  flex-wrap: nowrap;
+      flex-wrap: nowrap;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   -ms-flex-line-pack: center;
-  align-content: center;
+      align-content: center;
   -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
+      -ms-flex-align: center;
+          align-items: center;
   padding: 0 25px;
   margin-bottom: 80px;
 }
@@ -141,17 +138,17 @@ main {
   border-radius: 35px;
   -webkit-box-orient: horizontal;
   -webkit-box-direction: normal;
-  -ms-flex-direction: row;
-  flex-direction: row;
+      -ms-flex-direction: row;
+          flex-direction: row;
   -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
+      flex-wrap: wrap;
   text-align: center;
   -webkit-box-flex: 1;
-  -ms-flex-positive: 1;
-  flex-grow: 1;
+      -ms-flex-positive: 1;
+          flex-grow: 1;
   -webkit-box-ordinal-group: 4;
-  -ms-flex-order: 3;
-  order: 3;
+      -ms-flex-order: 3;
+          order: 3;
   padding: 25px;
   margin: 0 25px;
   width: 50%;
@@ -159,20 +156,20 @@ main {
 
 .game-container div.board-wrapper .token-wrapper {
   -webkit-box-flex: 1;
-  -ms-flex: 1 0 calc((100% / 3) - 10px);
-  flex: 1 0 calc((100% / 3) - 10px);
+      -ms-flex: 1 0 calc((100% / 3) - 10px);
+          flex: 1 0 calc((100% / 3) - 10px);
   height: 22.5vh;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
   -ms-flex-line-pack: center;
-  align-content: center;
+      align-content: center;
   -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
+      -ms-flex-align: center;
+          align-items: center;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   font-family: 'RammettoOne-Regular', 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
   font-size: 5rem;
   background-color: #111491;
@@ -198,7 +195,7 @@ main {
 
 .game-container div.board-wrapper .token-wrapper:hover::before {
   -webkit-transform: scale(1.8);
-  transform: scale(1.8);
+          transform: scale(1.8);
 }
 
 .game-container div.board-wrapper .player-1-cell.token-wrapper::before,
@@ -249,35 +246,35 @@ main {
 
 .game-container #player-1 {
   -webkit-box-ordinal-group: 2;
-  -ms-flex-order: 1;
-  order: 1;
+      -ms-flex-order: 1;
+          order: 1;
 }
 
 .game-container #player-2 {
   -webkit-box-ordinal-group: 4;
-  -ms-flex-order: 3;
-  order: 3;
+      -ms-flex-order: 3;
+          order: 3;
 }
 
 .game-container .player-card {
   -webkit-box-flex: 0.5;
-  -ms-flex-positive: 0.5;
-  flex-grow: 0.5;
+      -ms-flex-positive: 0.5;
+          flex-grow: 0.5;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
   -webkit-box-orient: vertical;
   -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
+      -ms-flex-direction: column;
+          flex-direction: column;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
+      -ms-flex-align: center;
+          align-items: center;
   -ms-flex-line-pack: center;
-  align-content: center;
+      align-content: center;
   margin: 25px;
   padding: 40px 0;
   width: 10vw;
@@ -316,21 +313,21 @@ main {
   display: -ms-flexbox;
   display: flex;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   -webkit-box-orient: vertical;
   -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
+      -ms-flex-direction: column;
+          flex-direction: column;
   -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
+      flex-wrap: wrap;
   width: 100%;
 }
 
 .start-screen div.game-title .row {
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -342,7 +339,7 @@ main {
   padding: 4px;
   font-size: 5.5rem;
   color: #ffd033;
-  text-shadow: 0 1px #2a1c52, -1px 0 #3b299b, -1px 2px #2a1c52, -2px 1px #3b299b, -2px 3px #2a1c52, -3px 2px #3b299b, -3px 4px #2a1c52, -4px 3px #3b299b, -4px 5px #2a1c52, -5px 4px #3b299b, -5px 6px #2a1c52, -6px 5px #3b299b, -6px 7px #2a1c52, -7px 6px #3b299b, -7px 8px #2a1c52, -8px 7px #3b299b;
+  text-shadow: 0 1px #2a1c52, -1px 0 #3b299b, -1px 2px #2a1c52, -2px 1px #3b299b, -2px 3px #2a1c52, -3px 2px #3b299b, -3px 4px #2a1c52, -4px 3px #3b299b, -4px 5px #2a1c52, -5px 4px #3b299b, -5px 6px #2a1c52, -6px 5px #3b299b, -6px 7px #2a1c52, -7px 6px #3b299b, -7px 8px #2a1c52, -8px 7px #3b299b;
 }
 
 .start-screen div.game-title .row-1 span:nth-child(even),
@@ -359,8 +356,8 @@ main {
   display: -ms-flexbox;
   display: flex;
   -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
+      -ms-flex-pack: center;
+          justify-content: center;
 }
 
 .buttons-container .new-game {
@@ -395,3 +392,4 @@ main {
   right: 10px;
   top: -20px;
 }
+/*# sourceMappingURL=styles.css.map */
