+++ assets/js/board.js
@@ -1,8 +1,91 @@
-/* eslint-disable import/extensions */
-/* eslint-disable import/no-cycle */
-import GameLogic from './game.js';
-
 const GameBoard = (() => {
+  
+  let movesArray = ['', '', '', '', '', '', '', '', ''];
+  const message = document.querySelector('#instructions');
+
+  const beginTurn = (currentPlayer) => {
+    message.innerHTML = `Hello ${currentPlayer.name}, this is your turn.<br>`;
+    message.innerHTML += '<span>Please, click on the cell you want to play.</span>';
+    document.getElementById('instructions').classList.toggle('information');
+    document.getElementById('instructions').classList.toggle('information');
+  };
+
+  const getWinnerMessage = (currentPlayer) => {
+    message.innerHTML = `We have a winner! Congratulations ${currentPlayer.name}.`;
+    document.getElementById('instructions').classList.toggle('winner');
+  }
+
+  const getTieMessage = () => {
+    message.innerHTML = 'We have a tie! Start a new game.';
+    document.getElementById('instructions').classList.toggle('information');
+  }
+
+  const disableCells = () => {
+    for (let i = 0; i <= 8; i += 1) {
+      const cell = document.getElementById(`${i}`);
+      cell.style.pointerEvents = 'none';
+    }
+  };
+
+  const togglePlayerClass = (currentPlayer, player, element) => {
+    if (currentPlayer === player) {
+      element.classList.toggle('player-1-cell');
+    } else {
+      element.classList.toggle('player-2-cell');
+    }
+  };
+
+  const getPlayerSymbol = (tokens) => {
+    let playerSymbol = prompt(`Choose your token from the following options: ${tokens.join(' ')}.`, '').toLowerCase();
+    while (!tokens.includes(playerSymbol)) {
+      playerSymbol = prompt(`Please select a valid option. Choose from: ${tokens.join(' ')}.`, '');
+    }
+
+    if (playerSymbol != null) {
+      alert(`Cool! ${playerSymbol}! Your token will be ${playerSymbol}`);
+    }
+
+    return playerSymbol;
+
+
+    // const player = Player(playerName, playerSymbol);
+    // return player;
+  };
+
+  const getPlayerName = () => {
+    let playerName = prompt('Please enter your name', '');
+    while (playerName === '') {
+      playerName = prompt('Name cannot be empty, please provide one', '');
+    }
+
+    if (playerName !== null) {
+      alert(`Hello ${playerName}! Welcome to your tic-tac-toe game`);
+    }
+
+    return playerName;
+  }
+
+  const playerInput = (i) => {
+    const currentCell = document.getElementById(`${i}`);
+    if (currentCell.innerHTML === '') {
+      currentCell.innerHTML = currentPlayer.token;
+      currentCell.classList.toggle('selected-cell');
+      togglePlayerClass(player1, currentCell);
+      movesArray[i] = currentPlayer.token;
+      evaluateWinner();
+    } else {
+      alert('This option has been taken. Please select an empty option.');
+    }
+  };
+
+  const toggleStartElements = () => {
+    const startScreen = document.getElementById('start-screen');
+    const newGameBtn = document.getElementById('new-game-btn');
+    startScreen.className += ' float-game-title';
+    newGameBtn.className += ' float-new-game-btn';
+  };
+
+
   const displayBoard = () => {
     const mainElement = document.getElementById('game-board');
     const container = document.createElement('div');
@@ -65,6 +148,8 @@ const GameBoard = (() => {
     deleteBoard,
     displayPlayers,
     deletePlayers,
+    getWinnerMessage,
+    getTieMessage
   };
 })();
