+++ assets/js/game.js
@@ -4,11 +4,14 @@ import Player from './player.js';
 import GameBoard from './board.js';
 
 const GameLogic = (() => {
-  let player1;
-  let player2;
+  
   let currentPlayer;
   let winner;
   let tokens = ['x', 'o', '#', '$', '@', '!'];
+  let player1;
+  let player2;
+
+  const message = document.querySelector('#instructions');  
   const winningPatterns = [
     [0, 1, 2],
     [3, 4, 5],
@@ -18,49 +21,7 @@ const GameLogic = (() => {
     [2, 5, 8],
     [0, 4, 8],
     [2, 4, 6]];
-  let movesArray = ['', '', '', '', '', '', '', '', ''];
-  const message = document.querySelector('#instructions');
-
-  const disableCells = () => {
-    for (let i = 0; i <= 8; i += 1) {
-      const cell = document.getElementById(`${i}`);
-      cell.style.pointerEvents = 'none';
-    }
-  };
-
-  const togglePlayerClass = (player, element) => {
-    if (currentPlayer === player) {
-      element.classList.toggle('player-1-cell');
-    } else {
-      element.classList.toggle('player-2-cell');
-    }
-  };
-
-  const getPlayerVariables = (tokens) => {
-    let playerName = prompt('Please enter your name', '');
-    while (playerName === '') {
-      playerName = prompt('Name cannot be empty, please provide one', '');
-    }
-
-    if (playerName !== null) {
-      alert(`Hello ${playerName}! Welcome to your tic-tac-toe game`);
-    }
-
-    let playerSymbol = prompt(`Choose your token from the following options: ${tokens.join(' ')}.`, '').toLowerCase();
-    while (!tokens.includes(playerSymbol)) {
-      playerSymbol = prompt(`Please select a valid option. Choose from: ${tokens.join(' ')}.`, '');
-    }
-
-    if (playerSymbol != null) {
-      alert(`Cool! ${playerSymbol}! Your token will be ${playerSymbol}`);
-    }
-
-    const index = tokens.indexOf(playerSymbol);
-    tokens.splice(index, 1);
-
-    const player = Player(playerName, playerSymbol);
-    return player;
-  };
+  
 
   const reset = () => {
     GameBoard.deleteBoard();
@@ -73,22 +34,22 @@ const GameLogic = (() => {
     currentPlayer = null;
   };
 
-  const beginTurn = () => {
-    message.innerHTML = `Hello ${currentPlayer.name}, this is your turn.<br>`;
-    message.innerHTML += '<span>Please, click on the cell you want to play.</span>';
-    document.getElementById('instructions').classList.toggle('information');
-    document.getElementById('instructions').classList.toggle('information');
-  };
-
   const start = () => {
     reset();
-    player1 = getPlayerVariables(tokens);
-    player2 = getPlayerVariables(tokens);
+    player1 = Player(GameBoard.getPlayerName, GameBoard.getPlayerSymbol(tokens));
+    tokens = updateTokens(player1.token, tokens);
+    player2 = Player(GameBoard.getPlayerName, GameBoard.getPlayerSymbol(tokens));
     currentPlayer = player1;
     beginTurn(currentPlayer);
     GameBoard.displayBoard();
     GameBoard.displayPlayers(player1, player2);
   };
+
+  const updateTokens = (selectedToken, tokensArr) => {
+    const index = tokensArr.indexOf(selectedToken);
+    tokensArr.splice(index, 1);
+    return tokensArr;
+  }
 
   const switchPlayer = () => {
     currentPlayer = currentPlayer === player1 ? player2 : player1;
@@ -96,13 +57,10 @@ const GameLogic = (() => {
 
   const endGame = (winner) => {
     if (winner) {
-      message.innerHTML = `We have a winner! Congratulations ${currentPlayer.name}.`;
-      document.getElementById('instructions').classList.toggle('winner');
       disableCells();
       return true;
     } if (movesArray.every(el => el !== '')) {
-      message.innerHTML = 'We have a tie! Start a new game.';
-      document.getElementById('instructions').classList.toggle('information');
+      
       disableCells();
       return true;
     }
@@ -125,30 +83,8 @@ const GameLogic = (() => {
     }
   };
 
-  const playerInput = (i) => {
-    const currentCell = document.getElementById(`${i}`);
-    if (currentCell.innerHTML === '') {
-      currentCell.innerHTML = currentPlayer.token;
-      currentCell.classList.toggle('selected-cell');
-      togglePlayerClass(player1, currentCell);
-      movesArray[i] = currentPlayer.token;
-      evaluateWinner();
-    } else {
-      alert('This option has been taken. Please select an empty option.');
-    }
-  };
-
-  const toggleStartElements = () => {
-    const startScreen = document.getElementById('start-screen');
-    const newGameBtn = document.getElementById('new-game-btn');
-    startScreen.className += ' float-game-title';
-    newGameBtn.className += ' float-new-game-btn';
-  };
-
   return {
-    start,
-    playerInput,
-    toggleStartElements,
+    start
   };
 })();
